requirements.txt
apache-airflow==2.7.0
apache-airflow-providers-postgres==2.0.0